// src/models/findAllPostsByUserIdModelModified.ts
export interface findAllPostsByUserIdResponseModified {
    isActive: boolean
    _id: string
    post: string
    userId: string
    userName: string
    userPhotoId: string
    postImageId: string
    isAdmin: boolean
    profession: string
    createdDate: string
    __v: number
    id: string
    postImageUrl: string
  }
  